msm: native: fix memory leak during snapshot
The JPEG buffers were being allocated from the MM ION heap during
snapshot. However, the Jpeg buffer fd wasn't being closed because
of which there was a memory leak. This bug crept in because of a
recent change that removed use of PMEM for JPEG memory allocation.

This is port of:
https://www.codeaurora.org/cgit/quic/la/platform/hardware/qcom/camera/commit/?h=kk_2.7_rb1.21&id=4ca6b47299da802a2975884559da4c8fed5ed76c