Camera: Fix for main image corruption
Use proper luma offset while registering the snapshot buffer with
kernel. In the present code luma offset is hard coded to 0 when
registering , this holds good only when the image height is a
multiple of 16.But, for the cases where the image height is not
a multiple of 16 and for the jpeg rotation values of 90 and 180,
the luma offset is calculated differently and it becomes a non
zero value.So, use this calculated luma offset value instead
of hard coded 0 when registering the buffer.

This is port of:
https://www.codeaurora.org/cgit/quic/la/platform/hardware/qcom/camera/commit/?h=kk_2.7_rb1.21&id=4777aa1f7fccfbf70ea5a303efa14e423c9efb40